+++ packages/core-js/stage/2.js
@@ -4,6 +4,7 @@ require('../proposals/iterator-helpers');
 require('../proposals/map-upsert-stage-2');
 require('../proposals/set-methods');
 require('../proposals/using-statement');
+require('../proposals/well-formed-unicode-strings');
 // TODO: Obsolete versions, remove from `core-js@4`
 require('../proposals/array-grouping');
 require('../proposals/decorators');